feat: core eksporterer to interfaces for children
React 18 fjerner children som implicit prop. Om komponenten din bare trenger children, eller du vil
slippe å deklarere den overalt, extend disse to typene.
